% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.decay.R
\name{beta.decay}
\alias{beta.decay}
\title{Fit an exponential distance decay model for beta diversity}
\usage{
beta.decay(gradient, counts, coords, nboots = 1000, dis.fun = "beta.pair",
  dis = 1, like.pairs = T, ...)
}
\arguments{
\item{gradient}{A vector of environmental values of interests}

\item{counts}{A dataframe of community matrix. Presence/absence if dis.fun="beta.pair"}

\item{coords}{A dataframe of 2 column with x and y coordinates of the sampling site. The
same sites will be remove from bootsrap procedures if like.pairs = TRUE}

\item{nboots}{A integer number indicating the numbers of bootsrap resampling}

\item{dis.fun}{Matching function to compute dissimilarity. The default is \code{\link[betapart]{beta.pair}}}

\item{dis}{An integer of 1, 2 or 3, indicating either the spatial turnover, nestness or total dissmilarity
are used in the distance decay modeling.}

\item{like.pairs}{Logical, whether to remove the like pairs or not. During bootstrap resampling,
because the sites are randomyl resampled with replacement and hence there may be multiple copies
of some sites. like.pairs = TRUE removes those site pairs having zero separation distance.}

\item{...}{Arguments to be passed to the matching function}
}
\value{
A list of data frames, including:
\describe{
  \item{Results}{A data frame of bootstrap estimate of GLM coefficients, (intercept) and x, as well as beta diversity at zero distance (s0) and halving distance (halfd)}
  \item{Summary}{A data frame of bootsrap summary statistics}
  \item{Predictions}{A data frame of distance difference, bootstrap predictions and 95\% confidence intervals}
  \item{CtrlList}{Other information}
}
}
\description{
Fit a generalized dissimilarity model to the spatial turnover, nestness or
dissmilarity of the ecological assemblages(including zero similarity points)
between two sites as a function of their distance apart along an environmental
gradient.
}
\details{
This funtion extends \code{\link{dist.decay}} to fit a generalized linear models (GLMs)
using quasibinomial distribution (with a log-link function) for the partition or total beta diversity
in an ecological assemblage as a function of an environmental gradient. The beta diversity is partitioned
into spatial turnover, nestness and totdal dissimilarity using \code{\link[betapart]{beta.pair}},
\code{\link[betapart]{bray.part}}, \code{\link[betapart]{functional.beta.pair}}, or \code{\link[betapart]{phylo.beta.pairt}}
following the methods decribed by Baselga (2010, 2012). A modified bootstrap method was used to estimate
the bootraped mean and standard error of model parameters (alpha, beta), \eqn{s=1-alpha*exp(-beta*d)} for
the spatial turnover and dissimilarity and \eqn{s=alpha*exp(-beta*d)} for the nestedness component of betat diversity,
where s is assemblage sptail turnover, nestedness or disimilarity and d is environmental distance,
and then use these parameters to calculate the turnover, nestedness or disimilarity at zero distance (s0)
and the halving distance (halfd) for which the turnover, nestedness or disimilarity between sites
decreased/increase by 50\% following the method decribed by Millar et al. (2010).
}
\note{
This function models the complement of dissimilarity (similarity).
}
\examples{
data(os)
dd <- beta.decay(gradient=os$dist, counts=decostand(os[, -1:-7], method="pa"),
                 coords=os[, c("longitude", "latitude")], nboots=1000,
                 dis.fun = "beta.pair", index.family = "sorensen", dis = 1, like.pairs=T)
x <- vegdist(os$dist, method = "euclidean")
y <- 1 - dist(decostand(os[,-1:-7], method="pa"), index.family = "sorensen")[[1]]
plot(x, y)
lines(dd$Predictions[, "x"], dd$Predictions[,"mean"], col="red", lwd=2)
}
\author{
Chih-Lin Wei <chihlinwei@gmail.com>
}
\references{
\itemize{
  \item Millar, R.B., Anderson, M.J., Tolimieri, N., 2011. Much ado about nothings: using zero similarity points in distance-decay curves. Ecology 92, 1717-1722. doi:10.1890/11-0029.1
  \item Baselga, A. 2010. Partitioning the turnover and nestedness components of beta diversity. Global Ecology and Biogeography 19:134-143
  \item Baselga, A. 2012. The relationship between species replacement, dissimilarity derived from nestedness, and nestedness. Global Ecology and Biogeography 21, 1223-1232
}
}
\seealso{
\code{\link{dist.decay}}
}

