% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.decay.R
\name{dist.decay}
\alias{dist.decay}
\title{Fit an exponential distance decay model to assemblage similarities}
\usage{
dist.decay(gradient, counts, coords, nboots = 1000, dis.fun = "vegdist",
  method = "bray", like.pairs = T)
}
\arguments{
\item{gradient}{A vector of environmental values of interests}

\item{counts}{A dataframe of community matrix of species abundance data}

\item{coords}{A dataframe of 2 column with x and y coordinates of the sampling site. The
same sites will be remove from bootsrap procedures if like.pairs = TRUE}

\item{nboots}{A integer number indicating the numbers of bootsrap resampling}

\item{dis.fun}{Matching function to compute dissimilarity. The default is \code{\link[vegan]{vegdist}}}

\item{method}{Dissimilarity index, partial match to "manhattan", "euclidean", "canberra",
"bray", "kulczynski", "jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup" ,
"binomial", "chao", "cao" or "mahalanobis".}

\item{like.pairs}{Logical, whether to remove the like pairs or not. During bootstrap resampling,
because the sites are randomly resampled with replacement and hence there may be multiple copies
of some sites. like.pairs = TRUE removes those site pairs having zero separation distance.}
}
\value{
A list of data frames, including:
\describe{
  \item{Results}{A data frame of bootstrap estimate of GLM coefficients, (intercept) and x, as well as beta diversity at zero distance (s0) and halving distance (halfd}
  \item{Summary}{A data frame of bootsrap summary statistics}
  \item{Predictions}{A data frame of distance difference, bootstrap predictions and 95\% confidence intervals}
  \item{CtrlList}{Other information}
}
}
\description{
Fit a generalized dissimilarity model in ecological similarity of assemblages
(including zero similarity points) between two sites as a function of their
distance apart along an environmental gradient.
}
\details{
This funtion fits a generalized linear models (GLMs) using quasibinomial distribution (with
a log-link function) for ecological similarity of assemblages between any two sites as a function
of their distance apart along a environmental gradient. A modified bootstrap method was used to estimate
the bootraped mean and standard error of model parameters (alpha, beta), \eqn{s=alpha*exp(-beta*d)},
where s is assemblage similarity and d is environmental distance, and then use these parameters to calculate
the similarity at zero distance (s0) and the halving distance (halfd) for which the similarity between sites
decreased by 50\% following the method decribed by Millar et al. (2010).
}
\note{
This function models the complement of dissimilarity (similarity).
}
\examples{
# Ostrocode exponential distance decay
data(os)
dd <- dist.decay(gradient=os$dist, counts=os[, -1:-7], coords=os[, c("longitude", "latitude")], nboots=1000, dis.fun = "vegdist", method = "bray", like.pairs=T)
x <- vegdist(os$dist, method = "euclidean")
y <- 1-vegdist(os[, -1:-7], method = "bray")
plot(x, y)
lines(dd$Predictions[, "x"], dd$Predictions[,"mean"], col="red", lwd=2)

}
\author{
Chih-Lin Wei <chihlinwei@gmail.com>
}
\references{
\itemize{
  \item Millar, R.B., Anderson, M.J., Tolimieri, N., 2011. Much ado about nothings: using zero similarity points in distance-decay curves. Ecology 92, 1717<U+FFFD><U+FFFD>1722. doi:10.1890/11-0029.1
}
}
\seealso{
\code{\link{beta.decay}}
}

